Method calcEnergyFluxMap included
The method calcEnergyFluxMap (and getEnergyFluxMap) was included in the heat class. It is used in the context of a two-temperature model, for example, as used when simulating electon-phonon coupling dynamics. The result gives the total energy flux (or change rate) in both subsystems due to subsystem coupling (e.g. electron-phonon coupling) and diffusion within each subsystem. Specifically, the method evaluates the individual terms in the two-temperature model using the tempMap calculated before.