Generalization of thermCond in heat.m
Modifications in pdeHeatEquation() to allow for using every subsystem temperature in each subsystem definition of thermCond. The analytical function can now use T(1),...,T(N) in each of the subsystems. This is in analogy to the usage in the property subsytemCoupling where the diference T(2)-T(1) has to be evaluated.